Fix incorrect calculation in subtask g) of task 1
The variances and standard deviations of the two datasets are now
calculated as samples (bodovy odhad) instead of as based on the entire
population (populacny odhad). This fixes the testing criterium as well,
as it uses the estimated sample variance result. These values are fixed
in the documentation as well.