Add methods for chaining to Employee class
The workOn() method checks the job of the current employee and if it
matches the requested job, calls the finish() method, which finishes the
job. Otherwise it calls workOn() of the next employee in the chain.
The finish() method in the Employee class is abstract and needs to
be implemented for each Employee subclass.