Begin building the algorithm for getting the Nt set
The "Nt" set is the set of nonterminals produced by algorithm 4.1 from
the TIN scripts. It is basically the set of non-terminals, which
generate something (so if in the input there is a non-terminal A with
the only rule of A -> aA, this does not generate anything, since it
would loop forever - therefore it would not be in "Nt").