RESET the project progress
I was not happy with the direction this project was heading, so I
completely revamped it. For this project there is no concrete
assignment, which means I have to create my own, and the previous
`Employee manager` or whatever it wanted to be, was not good. I think.
So I'm changing the direction entirely and beginning work on the second
idea, which is some sort of a `Todo manager`.

I am continuing in this branch and not scrapping previous progress for
reference/other purposes.