Modify the IC method for keylength calculation
The keylength is no longer the value which has it's IC closest to 0.065,
but rather the first one, that is within the interval <0.062,0.073>. The
upper bound shouldn't matter, but if the IC of a specific keylength isat
least 0.062, then the likelyhood of it being the correct keylength is
quite high.

The reason why it was changed from the absolute closest IC value to
0.065 is because, for some reason, multiples of the correct keylength
tended to have "better" IC values (i.e. closer to 0.065). I don't know
the reason behind this behaviour.