Near complete overhaul -- working state
Took the code and rewrote it basically from scratch. It is less
hardcoded to the expected 16 input values, more general, and it works.
However, it needs some cleanup. Mostly just removing stderr debug prints
and some DRY-ness, as some code is repeated. It may not be pretty to
look at, but it works and is fairly well commented.