Move repeated parts of code into a function
Value passing in merge() was terribly repetitive. This was remedied by
moving the repeated code to a new function pass_through() and adding a
control variable, which tells the new function whether the value should
be passed to the next process or to stdout (in the case of the last
process).