Simplify check for an edge case
If `from_top` + `from_bot` is equal to `2 * pow(2, rank-1) - 1`, then
that means, that either the top or bottom queues are full. This would
mean, that the check behind the currently deleted one would pass
anyways and the exact same value would be passed as in the deleted if
block. Therefore the deleted if block was unneeded, but an additional
queue length check needed to be added for correctly resetting the top
and bottom counters.