Add EOF encoding to Huffman tree
A new symbol was added for EOF. if this symbol is included in the code
stream, then upon reaching it during decoding, the decoder stops and
does not read the rest of the stream regardless of how much data is
still in the stream. The data read up to this point is decoded as
normal.