Add capability for vertical run-length encoding
Used sometimes when the `-a` option is passed. If the program deems it
better to encode in the vertical direction, then it encodes in the
vertical direction. The direction is selected based on how many changes
in pixel values occur in either direction - the fewer changes, the
better the direction.